sheez... another mistake I made in making changes for code review. ALWAYS TEST!

+++ src/hg/hgTracks/simpleTracks.c
@@ -3274,8 +3274,8 @@ for (item = items; item; item = item->next)
     if (positiveRangeIntersection(start, end, winStart, winEnd) <= 0)
 	continue;
 
-    int x1 = max(start - winStart, 0); 
-    int x2 = min(end - winStart, size);
+    int x1 = max((int)start - (int)winStart, 0); 
+    int x2 = min((int)end - (int)winStart, size);
 
     for(; x1 < x2; x1++)
 	counts[x1]++;
